Move pull and push features to ListActivity
