+++ client/app/src/main/java/com/example/client/MainActivity.java
@@ -20,20 +20,13 @@ import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.example.client.tflite.Classifier;
-import com.example.client.tflite.Classifier.Device;
-import com.example.client.tflite.Classifier.Model;
-import com.example.client.tflite.HttpTask;
 
-import org.json.JSONException;
-import org.json.JSONObject;
-
-import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.nio.channels.FileChannel;
 import java.util.List;
-import java.util.concurrent.ExecutionException;
+
+import static com.example.client.ListActivity.classifier;
 
 
 public class MainActivity extends AppCompatActivity {
@@ -41,11 +34,6 @@ public class MainActivity extends AppCompatActivity {
     protected final String tag = MainActivity.class.getSimpleName();
 
     private static Context context;
-
-    private Model model = Model.FLOAT_MOBILENET;
-    private Device device = Device.CPU;
-    private int numThreads = -1;
-    private Classifier classifier;
 
     private Bitmap image;
     private Integer imageId;
@@ -71,8 +59,8 @@ public class MainActivity extends AppCompatActivity {
 
         Intent intent = getIntent();
         imageId = intent.getIntExtra("position", R.drawable.demo01);
-
-        createClassifier(model, device, numThreads);
+        imageSizeX = intent.getIntExtra("imageSizeX", 0);
+        imageSizeY = intent.getIntExtra("imageSizeY", 0);
 
         /* Load image */
         imageView = (ImageView)findViewById(R.id.demo_image);
@@ -138,38 +126,20 @@ public class MainActivity extends AppCompatActivity {
                 Toast toast;
                 if (labelIndex == -1) {
                     toast = Toast.makeText(getApplicationContext(), String.format("(%s) is not a valid label :(", feedbackInput), Toast.LENGTH_SHORT);
+                    toast.show();
                 } else {
                     saveLabel(imageId, labelIndex);
                     toast = Toast.makeText(getApplicationContext(), String.format("(%s) Thanks for your feedback :)", feedbackInput), Toast.LENGTH_SHORT);
                     Log.d(tag, String.format("Correct output: %s", feedbackInput));
+                    toast.show();
+                    finish();
                 }
-                toast.show();
-                finish();
             }
         });
         builder.setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
             @Override
             public void onClick(DialogInterface dialog, int which) {
                 dialog.cancel();
-            }
-        });
- 
-        /* Pull & push */
-        button = (Button) findViewById(R.id.pull_model_params);
-        button.setOnClickListener(new Button.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Log.d(tag, "Update button triggered");
-                pullUpdatedParams();
-                createClassifier(model, device, numThreads);
-            }
-        });
-        button = (Button) findViewById(R.id.push_features);
-        button.setOnClickListener(new Button.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Log.d(tag, "Push button triggered");
-                pushIntermediateFeature();
             }
         });
     }
@@ -192,66 +162,6 @@ public class MainActivity extends AppCompatActivity {
         wrong.setVisibility(View.VISIBLE);
     }
 
-    /* Initializing Classifier */
-    private void createClassifier(Model model, Device device, int numThreads) {
-        if (classifier != null) {
-            Log.d(tag, "Closing classifier");
-            classifier.close();
-            classifier = null;
-        }
-        try {
-            Log.d(tag, "Creating classifier");
-            classifier = Classifier.create(this, model, device, numThreads);
-        } catch (IOException e) {
-            Log.e(tag, Log.getStackTraceString(e));
-        }
-
-        imageSizeX = classifier.getImageSizeX();
-        imageSizeY = classifier.getImageSizeY();
-    }
-
-    private void pullUpdatedParams() {
-        try {
-            String response = new HttpTask().execute("get").get();
-            if (response != null) {
-                classifier.close();
-                FileOutputStream fos = openFileOutput(classifier.getModelPath(), MODE_PRIVATE);
-                byte[] content = response.getBytes("ISO-8859-1");
-                fos.write(content);
-                fos.flush();
-                fos.close();
-            }
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        } catch (ExecutionException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    private void pushIntermediateFeature() {
-        try {
-            File files_dir = getFilesDir();
-            for (String filename: files_dir.list()) {
-                if (filename.startsWith("intermediates") || filename.startsWith("label")) {
-                    File file = new File(files_dir, filename);
-                    String response = new HttpTask().execute("post", file.getAbsolutePath()).get();
-                    if (response != null) {
-                        Log.d(tag, "HTTP Response: " + response);
-                        JSONObject json = new JSONObject(response);
-                        Log.d(tag, "success: " + json.getString("success"));
-                    }
-                }
-            }
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        } catch (ExecutionException e) {
-            e.printStackTrace();
-        } catch (JSONException e) {
-            e.printStackTrace();
-        }
-    }
 
     private void saveLabel(int imageId, int labelIndex) {
         String filename = "label_" + imageId;
