+++ client/app/src/main/res/layout/activity_main.xml
@@ -42,30 +42,6 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.0" />
 
-    <Button
-        android:id="@+id/push_features"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Push Intermediate Features"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintHorizontal_bias="0.498"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.092" />
-
-    <Button
-        android:id="@+id/pull_model_params"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Pull Updated Model"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintHorizontal_bias="0.498"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.162" />
-
     <ImageView
         android:id="@+id/demo_image"
         android:layout_width="fill_parent"
